Implement City's CRUD to pass the tests

+++ bds02/src/main/java/com/devsuperior/bds02/services/CityService.java
@@ -7,14 +7,12 @@ import com.devsuperior.bds02.services.exceptions.DatabaseException;
 import com.devsuperior.bds02.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class CityService {
@@ -24,26 +22,25 @@ public class CityService {
 
     @Transactional(readOnly = true)
     public List<CityDTO> findAll() {
-        List<City> list = cityRepository.findAll(Sort.by("name"));
-        return list.stream().map(CityDTO::new).collect(Collectors.toList());
+        return cityRepository.findAll(Sort.by("name")).stream().map(CityDTO::new).toList();
     }
 
     @Transactional
-    public CityDTO insert(CityDTO cityDTO) {
+    public CityDTO insert(CityDTO dto) {
         City city = new City();
-        city.setName(cityDTO.getName());
+        city.setName(dto.name());
 
         city = cityRepository.save(city);
-
         return new CityDTO(city);
     }
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!cityRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
         try {
             cityRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new ResourceNotFoundException("Recurso não encontrado");
         } catch (DataIntegrityViolationException e) {
             throw new DatabaseException("Falha de integridade referencial");
         }
